+++ CHANGELOG.md
@@ -10,6 +10,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -
 
+## [0.6.1] - 2024-01-29
+
+### Changed in 0.6.1
+
+- Update dependencies
+  - github.com/senzing-garage/g2-sdk-go v0.9.0
+  - github.com/senzing-garage/g2-sdk-go-base v0.5.0
+  - github.com/senzing-garage/g2-sdk-proto/go v0.0.0-20240126210601-d02d3beb81d4
+  - google.golang.org/grpc v1.61.0
+
 ## [0.6.0] - 2024-01-03
 
 ### Changed in 0.6.0